chore: add focus and connection update files
