+++ update_form.go
@@ -1,0 +1,46 @@
+package main
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// updateForm handles the add/edit connection form.
+func (m model) updateForm(msg tea.Msg) (model, tea.Cmd) {
+	if m.connections.form == nil {
+		return m, nil
+	}
+	var cmd tea.Cmd
+	switch msg.(type) {
+	case tea.WindowSizeMsg, tea.MouseMsg:
+		m.connections.manager.ConnectionsList, _ = m.connections.manager.ConnectionsList.Update(msg)
+	}
+	switch msg := msg.(type) {
+	case tea.KeyMsg:
+		switch msg.String() {
+		case "ctrl+d":
+			return m, tea.Quit
+		case "esc":
+			cmd := m.setMode(modeConnections)
+			m.connections.form = nil
+			return m, cmd
+		case "enter":
+			p, err := m.connections.form.Profile()
+			if err != nil {
+				if m.connections.statusChan != nil {
+					m.connections.statusChan <- err.Error()
+				}
+				return m, listenStatus(m.connections.statusChan)
+			}
+			if m.connections.form.index >= 0 {
+				m.connections.manager.EditConnection(m.connections.form.index, p)
+			} else {
+				m.connections.manager.AddConnection(p)
+			}
+			m.refreshConnectionItems()
+			cmd := m.setMode(modeConnections)
+			m.connections.form = nil
+			return m, cmd
+		}
+	}
+	f, cmd := m.connections.form.Update(msg)
+	m.connections.form = &f
+	return m, tea.Batch(cmd, listenStatus(m.connections.statusChan))
+}
